trx/scheduler: Use integer math for TOA (Timing of Arrival)
There's no need to express TOA as a float:
* We receive it as signed 16bit integer in units 1/256 symbol periods
* We pass it to L1SAP as signed integer in 1/4 symbol periods

So turn it into an int16_t with 1/256 symbol period accuracy throughout
the code to avoid both float arithmetic as well as loosing any precision.

+++ src/osmo-bts-trx/loops.h
@@ -14,7 +14,7 @@
  */
 
 int trx_loop_sacch_input(struct l1sched_trx *l1t, uint8_t chan_nr,
-	struct l1sched_chan_state *chan_state, int8_t rssi, float toa);
+	struct l1sched_chan_state *chan_state, int8_t rssi, int16_t toa);
 
 int trx_loop_sacch_clock(struct l1sched_trx *l1t, uint8_t chan_nr,
         struct l1sched_chan_state *chan_state);
